==> 1_analysis/0_descriptive_results.Rmd <==
---
title: "Descriptive Analyses"
output: html_document
date: "2022-08-23"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Matrix)
library(dbplyr)
library(tidyverse)
library(lubridate)
library(lme4)
library(mice)
library(gghalves)
library(emmeans)
library(naniar)
library(broom.mixed)
library(lmerTest)
library(scales)
library(ggdist)
library(scipub)
library(sjPlot)
library(brms)
library(psych)
beta_prior = c(prior_string("student_t(3, 0, 10)", class = "b"))
```

# Load data files

```{r}
load('../../cleaned_data/cleaned_data_for_daily_models.rda')
load('../../cleaned_data/cleaned_baseline_data.rda')
load('../../../../Wrangling/Self_Report/data/self_report_02_07_2024.Rda')
```


# Get # of participants at each visit

```{r}
total_n_participants = 179

# #s at each time point
selfreport_long %>%
  group_by(month) %>%
  distinct(ID) %>%
  group_by(month) %>%
  summarise(count = n(), pct = count/179)

```

# Join daily data with baseline 

```{r}
daily = left_join(daily, self_report_baseline, by = 'ID')

# Remove days after 200 days, participants who haven't finished study
daily = dplyr::filter(daily, day_num <= 200, TIMEPNT==6) %>%
    mutate(missing_all_passive = missing_gps +  missing_accel + missing_keyboard == 3, TRUE, FALSE) # Daily variable for whether all passive smartphone measures are missing


sum(daily$day_num >200)

days_in_study = daily %>%
  group_by(ID) %>%
  summarise(max_day_num = max(day_num))
```


# Get # of participant with current SI 
```{r}
current_si_check = self_report_baseline %>%
  dplyr::filter(ID %in% daily$ID) %>%
  mutate(current_si = SSI5_i > 0)


current_si_check %>%
  group_by(current_si) %>%
  count()
```


# Bring in weekly SI survey data
```{r}
load('../../cleaned_data/risk.rda')

# code weeks when risk surveys are missing
risk_missing = risk_week_present %>%
  dplyr::filter(week_num <= 28, days ==7) %>%  # only full weeks 
  group_by(ID) %>%
  mutate(personal_time_z = scale(week_num),
         summerbreak = month %in% c(7, 8),
         site = ifelse(startsWith(ID, '1'), 'CUIMC', 'UPMC'),
         siteCUMC = site == 'CUIMC') %>%
  ungroup() %>%
  left_join(., self_report_baseline, by = 'ID') 
```

# Service use descriptives
```{r}
self_report_baseline %>%
  dplyr::filter(ID %in% daily$ID) %>%
  dplyr::select(ID, current_medication, current_therapy, hospitalization_pastyear, visit_er_pastyear) %>%
  pivot_longer(-ID) %>%
  group_by(name, value) %>%
  summarize(n = n(), pct = n/179)
```

# More descriptive missingness plots

```{r}
dplyr::select(daily, daily, n_capture_day, val_sleep_duration_min, n_char_day) %>%
  naniar::vis_miss()

dplyr::select(daily, `Mood Survey`=daily, `GPS`=n_capture_day, `Accel`=val_sleep_duration_min, `Keyboard`=n_char_day) %>%
  naniar::gg_miss_upset()
```

# Generate table of missing data for baseline/self-report measures
```{r}
self_report_missing_df = self_report_baseline %>% 
  dplyr::filter(ID %in% daily$ID) %>%
  mutate(site = ifelse(startsWith(ID, '1'), 'CUIMC', 'UPMC'),
         siteCUMC = site == 'CUIMC') %>%
  dplyr::select(c(all_predictors, income_ordinal)) 

selfreport_missing_table = self_report_missing_df %>%
  pivot_longer(-ID) %>%
  group_by(name) %>%
  summarise(n_missing = sum(is.na(value)), pct_missing = round(n_missing/n()*100, 2))

time_varying_missing = daily %>% 
  mutate(val_sleep_duration_min = as.numeric(val_sleep_duration_min)) %>%
  dplyr::select(ID, daily, n_capture_day, val_sleep_duration_min, n_char_day) %>%
  pivot_longer(-ID) %>%
  group_by(name) %>%
  summarise(n_missing = sum(is.na(value)), pct_missing = round(n_missing/n()*100, 2))

total_days = nrow(daily)

risk_missing_summary = risk_missing %>%
  summarise(name = 'risk',
            n_missing = sum(risk_missing), pct_missing= n_missing/n()) 

missingness_table = 
  rbind(selfreport_missing_table, time_varying_missing, risk_missing_summary)

write.csv(missingness_table,file = '../../tables/missingness_pct.csv')
```

# Descriptive stats on missing data for different modalities

```{r}
daily_days_missing = daily %>%
  dplyr::filter(missing_daily==1, day_num < 200) 


daily_days_missing %>%
  group_by(missing_gps) %>%
  count() %>%
  mutate(prop = n/nrow(daily_days_missing))

daily_days_missing %>%
  group_by(missing_all_passive) %>%
  count() %>%
  mutate(prop = n/nrow(daily_days_missing))


load('../../cleaned_data/risk.rda')

risk_weekly %>% dplyr::filter(!is.na(riskOne)) %>% group_by(dayofweek) %>% summarise(n = n(), p = n/nrow(.))

risk_week_present = dplyr::select(risk_week_present, date=week, everything())

risk_week_present = mutate(risk_week_present, day = wday(date, label=TRUE))

risk_week_present = left_join(risk_week_present, dplyr::select(daily, date, ID, contains('missing')), by = c('date', 'ID'))

risk_days_missing = risk_week_present %>%
  dplyr::filter(risk_missing==1, week_num <= 28, TIMEPNT == 6) %>%
  dplyr::filter(!is.na(missing_daily))

risk_days_missing %>%
  group_by(missing_gps) %>%
  count() %>%
  mutate(prop = n/nrow(risk_days_missing))

risk_days_missing %>%
  mutate(missing_all_passive = missing_gps +  missing_accel + missing_keyboard == 3, TRUE, FALSE) %>%
  group_by(missing_all_passive) %>%
  count() %>%
  mutate(prop = n/nrow(risk_days_missing))
```

# EMA missingness ~ passive data missingness

Here, likelihood of missing EMA data is modeled as a function of whether ANY passive mobile sensor data was recorded the same day

```{r}
risk_week_present = mutate(risk_week_present, missing_all_passive = missing_gps +  missing_accel + missing_keyboard == 3, TRUE, FALSE)


risk_missing_passive_model = brms::brm(data = dplyr::filter(risk_week_present, week_num < 28), risk_missing ~ missing_all_passive + (1|ID), 
                                         family = bernoulli(link = 'logit'), cores = 4, chains = 4, iter = 2000, seed = 11291993, prior = beta_prior) 
daily_missing_passive_model = brms::brm(data = dplyr::filter(daily, day_num < 200), missing_daily ~ missing_all_passive + (1|ID), 
                                          family = bernoulli(link = 'logit'), cores = 4, chains = 4, iter = 2000, seed = 11291993, prior = beta_prior)

risk_missing_passive_model %>% emmeans(specs = 'missing_all_passive', type = 'response') %>%
  pairs(reverse=TRUE) %>%
  confint()

daily_missing_passive_model %>% emmeans(specs = 'missing_all_passive', type = 'response') %>%
  pairs(reverse=TRUE) %>%
  confint()

daily_missing_passive_model %>% emmeans(specs = 'missing_all_passive', type = 'response')

# check MLM outputs are consistent with raw data summaries (should differ a little)
daily %>%
  group_by(missing_all_passive) %>%
  summarise(n = n(), prop_missing_daily = sum(missing_daily)/n())

risk_week_present %>%
  mutate(missing_all_passive = missing_gps +  missing_accel + missing_keyboard == 3, TRUE, FALSE) %>%
  group_by(missing_all_passive) %>%
  summarise(n = n(), prop_missing_risk = sum(risk_missing)/n())
```


# Discriptive Tables on Study Cohort

```{r}
self_report2 = dplyr::filter(self_report_baseline, ID %in% daily$ID) %>%
  mutate(site = ifelse(startsWith(ID, '1'), 'NYC', 'Pittsburgh'))


cohort_table = scipub::FullTable1(dplyr::select(self_report2, sex, interview_age_years, site, White, Black, Asian, Multiracial, hispanic, hispanic, SITBI_sa_ever, SITBI_si_ever,  income_categorical, 
                                                MDDcurrent, AnxietyCurrent, SUDever, SITBI_engaged_nssi_ever))

sjPlot::tab_df(cohort_table$table, title='Table 1.', file = '../../tables/cohort_table1.doc')

# more descriptive summaries on cohort characteristics
summary(selfreport$income_ordinal)
table(selfreport$income_categorical)
summary(selfreport$comorbidity_current)


table(selfreport$heterosexual)
table(selfreport$cisgender)
```

# Supplemental Heatmap Figure Showing Participant Proportions of Misisngness for each Modality

```{r}
daily_pct_missing = daily %>%
  dplyr::select(ID, contains('missing'), -missing_all_passive) %>%
  pivot_longer(contains('missing')) %>%
  group_by(ID, name) %>%
  summarise(pct_missing = sum(value)/n())

risk_pct_missing = risk_missing %>%
  dplyr::select(ID, risk_missing) %>%
  group_by(ID) %>%
  summarise(pct_missing = sum(risk_missing)/n()) %>%
  mutate(name = 'risk')

pct_missing_combined = rbind(daily_pct_missing, risk_pct_missing) 


pct_missing_combined = dplyr::filter(pct_missing_combined, ID %in% self_report_baseline$ID,
                       ID %in% daily$ID,
                       ID %in% risk_missing$ID)

pct_missing_combined = pct_missing_combined %>%
  group_by(ID) %>%
  mutate(mean_pct = mean(pct_missing)) %>%
  ungroup()

pct_missing_combined = mutate(pct_missing_combined, 
            name = dplyr::recode(name, 'risk'='Weekly Suicidal Ideation Survey', 
                                 'missing_keyboard'=' Keyboard Input',
                                 'missing_gps'=' GPS',
                                 'missing_daily'='Daily Mood Survey',
                                 'missing_accel'=' Accelerometer'))

participants_descriptive_missing_all_datatypes_plt = pct_missing_combined %>%
  ggplot(data = ., aes(x = forcats::fct_reorder(ID, mean_pct), y = name, fill =pct_missing)) +
  geom_tile() + 
  scale_fill_viridis_c() +
  theme_classic() +
  labs(fill = 'Proportion Missing', x = 'Participants', y = NULL) +
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) 


ggsave(participants_descriptive_missing_all_datatypes_plt, filename = '../../figures/supplement/descriptive_all_missing_by_participant.png', height = 2, width = 10)
```

# Sexual Orientation

demooo_11 is sexual orientation

"1= Heterosexual
2= Gay or lesbian
3= Bisexual or pansexual
4= Asexual
5= Not sure
6= Other
99= Unknown or not reported"
```{r}
nda_demog = read.csv('/Volumes/auerbachlab/Columbia/MAPS_Data/NDA/July_2023/Existing_Structures/Interviews/DataScientists_DONE/demo02.csv', skip=1)


table(nda_demog$demooo_11)

non_na = sum(nda_demog$demooo_11 %in% 1:6)

table(nda_demog$demooo_11[nda_demog$demooo_11 %in% 1:6]) / non_na
```

# Autocorrelation of missingness (odds ratios)
```{r}
daily_for_autocor = daily %>%
  group_by(ID) %>%
  mutate(missing_daily_lag1 = lag(missing_daily, n =1),
         missing_keyboard_lag1 = lag(missing_keyboard, n =1),
         missing_gps_lag1 = lag(missing_gps, n =1),
         missing_accel_lag1 = lag(missing_accel, n =1)) %>%
  ungroup() %>%
  dplyr::select(ID, date, contains('missing'))


daily_autocor = lme4::glmer(data = daily_for_autocor, missing_daily ~ missing_daily_lag1 + (1|ID),
                            family = binomial(link = 'logit'))

keyboard_autocor = lme4::glmer(data = daily_for_autocor, missing_keyboard ~ missing_keyboard_lag1 + (1|ID),
                            family = binomial(link = 'logit'))

gps_autocor = lme4::glmer(data = daily_for_autocor, missing_gps ~ missing_gps_lag1 + (1|ID),
                            family = binomial(link = 'logit'))

accel_autocor = lme4::glmer(data = daily_for_autocor, missing_accel ~ missing_accel_lag1 + (1|ID),
                            family = binomial(link = 'logit'))


daily_autocor_or = broom.mixed::tidy(daily_autocor, exponentiate=TRUE, conf.int=TRUE) %>% mutate(type = 'daily')
keyboard_autocor_or = broom.mixed::tidy(keyboard_autocor, exponentiate=TRUE, conf.int=TRUE) %>% mutate(type = 'keyboard')
gps_autocor_or = broom.mixed::tidy(gps_autocor, exponentiate=TRUE, conf.int=TRUE) %>% mutate(type = 'gps')
accel_autocor_or = broom.mixed::tidy(accel_autocor, exponentiate=TRUE, conf.int=TRUE) %>% mutate(type = 'accel')

autocor_or_combined = rbind(daily_autocor_or, keyboard_autocor_or, gps_autocor_or, accel_autocor_or) %>%
  dplyr::filter(grepl('missing', term))

autocor_or_combined
```
