==> 1_analysis/2g_baseline_predictors_supplemental_figures.Rmd <==
---
title: "Make supplemental figures for baseline between-person predictors of missing data"
output: html_document
date: "2022-08-23"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Matrix)
library(dbplyr)
library(tidyverse)
library(lubridate)
library(lme4)
library(mice)
library(gghalves)
library(emmeans)
library(naniar)
library(broom.mixed)
library(lmerTest)
library(scales)
library(brms)
library(cowplot)
#library(ggfortify)
theme_set(theme_bw())
source('model_wrapper_functions.R')
```

# Load models

```{r}
load('../../cleaned_data/cleaned_data_for_daily_models.rda')
load('../../cleaned_data/cleaned_baseline_data.rda')

m = length(clinical_predictors)
corrected_conf_level = 1-(0.05/m)
```

# Function to make conditional effects (marginal predictions) plot
```{r}
make_conditional_effects_baseline = function(input_data, type){
    conditions = data.frame(personal_time_z = -1)
    input_data = input_data %>%
        mutate(cond_effects = map(model, ~brms::conditional_effects(., effects = 'scaled_predictor', conditions = conditions)))
    
    cond_effects_df = input_data$cond_effects[[1]]$scaled_predictor
    cond_effects_df$name = input_data$name[1]
    
    for (i in 2:nrow(input_data)){
        tmp_df = input_data$cond_effects[[i]]$scaled_predictor
        tmp_df$name = input_data$name[i]
        cond_effects_df = rbind(cond_effects_df, tmp_df)
    }
    
    cond_effects_df = mutate(cond_effects_df,
                             namefix = dplyr::recode(name, 'interview_age_years'='Age (1 Year)*',
                                    'SITBI_engaged_nssi_ever'='Lifetime NSSI (SITBI)',
                                    'MFQ_i'='Depression Symptoms (MFQ)',
                                    'SITBI_sa_ever'='Liftime Suicide Attempt (SITBI)', 'SHAPS_TOT_r_i'='Anhedonia (SHAPS)','comorbidity_current'='# Current Diagnoses',
                                    'SITBI_si_ever'='Lifetime Suicidal Ideaction (SITBI)', 'hospitalization_pastyear' = 'Past Year Hospitalization',
                                    'visit_er_pastyear'='Past Year ED Visit', 'SITBI_SB_ever'='Lifetime Suicidal Behavior (SITBI)',
                                    'SCARED_i'='Anxiety Symptoms (SCARED)', 'SSI19_i'='Suicidal Ideation Severity (SSI)',
                                    'INQ_burdenensomeness_i'='Percieved Burdensomeness (INQ)', 'INQ_thwartedbelonging_i'='Thwarted Belonging (INQ)',
                                    'hispanic'='Hispanic','income_ordinal'='Parental Income',
                                    'cisgender'='Cisgender', 'sexMTRUE'='Assigned Male Sex', 'genderMTRUE'='Male Gender', 'heterosexual'='Heterosexual',
                                    'siteCUMCTRUE'='Site (NYC)', 'parentscollege'='Parent Completed College', 'current_therapy'='Current Therapy',
                                    'current_medication'='Current Medication',
                                    'interview_date_numeric'='Interview Date*'))
    
    clinical_predictors_continuous = c('SHAPS_TOT_r_i', 'MFQ_i',
                     'SCARED_i', 'SSI19_i', 'INQ_burdenensomeness_i', 'INQ_thwartedbelonging_i',  'comorbidity_current')
    
    clinical_predictors_binary = c('current_medication', 'current_therapy', 'hospitalization_pastyear', 'visit_er_pastyear',
                                   'SITBI_SB_ever', 'SITBI_sa_ever','SITBI_engaged_nssi_ever', 'SITBI_si_ever')
    
    
    cond_effects_continuous = dplyr::filter(cond_effects_df, name %in% clinical_predictors_continuous)
    cond_effects_binary = dplyr::filter(cond_effects_df, name %in% clinical_predictors_binary)
    
    
    cond1 = cond_effects_continuous %>%
        ggplot(data = ., aes(x = scaled_predictor, y = estimate__)) +
            geom_ribbon(aes(ymin = lower__, ymax = upper__), alpha = .25)+
            geom_line() +
            facet_wrap('namefix') +
            labs(x = NULL, y = 'P(Missing)', title = type) 
    
    cond2 = cond_effects_binary %>%
        dplyr::filter(scaled_predictor %in% c(0,1)) %>%
        dplyr::mutate(yn = ifelse(scaled_predictor==1, 'Yes', 'No')) %>%
        ggplot(data = ., aes(x = yn, y = estimate__)) +
            geom_point() + 
            geom_errorbar(aes(ymin = lower__, ymax = upper__), width = 0.1)+
            facet_wrap('namefix') +
            labs(x = NULL, y = 'P(Missing)', title = '')
    
    
    pgrid = cowplot::plot_grid(cond1, cond2)
    
    return(pgrid)
}
```

# Conditional FX plots

## GPS
```{r}
load('model_outputs/baseline_predictors_missingness_brm_gps.rda')
baseline_conditional_fx_gps = make_conditional_effects_baseline(input_data = gps_outputs$models_df, type = 'Baseline Predictors of GPS Missingness')
cowplot::save_plot(baseline_conditional_fx_gps, filename = '../../figures/supplement/baseline_conditional_fx_gps.png', base_height = 7, base_width = 14)
rm(baseline_conditional_fx_gps, gps_outputs)
gc()
```

## Accel
```{r}
load('model_outputs/baseline_predictors_missingness_brm_accel.rda')
baseline_conditional_fx_accel = make_conditional_effects_baseline(input_data = accel_outputs$models_df, type = 'Baseline Predictors of Accelerometer Missingness')
cowplot::save_plot(baseline_conditional_fx_accel, filename = '../../figures/supplement/baseline_conditional_fx_accel.png', base_height = 7, base_width = 14)
rm(baseline_conditional_fx_accel, accel_outputs)
gc()
```

## Keyboard
```{r}
load('model_outputs/baseline_predictors_missingness_brm_keyboard.rda')
baseline_conditional_fx_keyboard = make_conditional_effects_baseline(input_data = keyboard_outputs$models_df, type = 'Baseline Predictors of Keyboard Input Missingness')
cowplot::save_plot(baseline_conditional_fx_keyboard, filename = '../../figures/supplement/baseline_conditional_fx_keyboard.png', base_height = 7, base_width = 14)
rm(baseline_conditional_fx_keyboard, keyboard_outputs)
gc()
```

## Daily
```{r}
load('model_outputs/baseline_predictors_missingness_brm_daily.rda')
baseline_conditional_fx_daily = make_conditional_effects_baseline(input_data = daily_outputs$models_df, type = 'Baseline Predictors of Daily Mood Survey Missingness')
cowplot::save_plot(baseline_conditional_fx_daily, filename = '../../figures/supplement/baseline_conditional_fx_daily.png', base_height = 7, base_width = 14)
rm(baseline_conditional_fx_daily, daily_outputs)
gc()
```

## SI Risk Survey
```{r}
load('model_outputs/baseline_predictors_missingness_brm_risk.rda')
baseline_conditional_fx_risk = make_conditional_effects_baseline(input_data = risk_outputs$models_df, type = 'Baseline Predictors of Weekly Suicidal Ideation Survey Missingness')
cowplot::save_plot(baseline_conditional_fx_risk, filename = '../../figures/supplement/baseline_conditional_fx_risk.png', base_height = 7, base_width = 14)
rm(baseline_conditional_fx_risk, risk_outputs)
gc()
```
