==> 1_analysis/2i_supplemental_daily_risk_between_mlm_subgroup.Rmd <==
---
title: "Model Between-Participant Baseline Predictors of Missing Data: Supplemental Analysis With School Closure Covariate"
output: html_document
date: "2022-08-23"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Matrix)
library(dbplyr)
library(tidyverse)
library(lubridate)
library(lme4)
library(mice)
library(gghalves)
library(emmeans)
library(naniar)
library(broom.mixed)
library(lmerTest)
library(scales)
library(brms)
#library(ggfortify)
theme_set(theme_bw())
source('model_wrapper_functions.R')
options(buildtools.check = function(action) TRUE)
```

# Data setup

```{r}
load('../../cleaned_data/cleaned_data_for_daily_models.rda')
load('../../cleaned_data/cleaned_baseline_data.rda')
```

```{r}
m = length(clinical_predictors)
corrected_conf_level = 1- (0.05/m)
```

```{r}
daily = left_join(daily, self_report_baseline, by = 'ID')
daily = dplyr::filter(daily, day_num <= 200)
daily = dplyr::filter(daily, TIMEPNT==6)
```

```{r}
load('../../../cleaned_data/covid.Rda')
covid = dplyr::select(covid, date, schoolclosure, stayathome, site) 
covid$schoolclosure = as.numeric(as.character(covid$schoolclosure))
covid = mutate(covid, site = ifelse(site=='PITT', 'UPMC', site))
daily = left_join(daily, covid, by = c('date', 'site'))
daily = mutate(daily, schoolclosure = ifelse(date > '2022-01-01', 0, schoolclosure))


daily %>% dplyr::filter(Control == 1) %>%
  group_by(ID) %>%
  count()
```

## Weekly risk data
```{r}
load('../../cleaned_data/risk.rda')


risk_missing = risk_week_present %>%
  dplyr::filter(week_num <= 28, days ==7) %>%
  group_by(ID) %>%
  mutate(personal_time_z = scale(week_num),
         summerbreak = month %in% c(7, 8),
         site = ifelse(startsWith(ID, '1'), 'CUIMC', 'UPMC'),
         siteCUMC = site == 'CUIMC') %>%
  ungroup() %>%
  left_join(., self_report_baseline, by = 'ID') 

risk_missing = dplyr::filter(risk_missing, TIMEPNT==6)

risk_missing$date = risk_missing$week
risk_missing = left_join(risk_missing, covid, by = c('date', 'site'))
risk_missing = mutate(risk_missing, schoolclosure = ifelse(date > '2022-01-01', 0, schoolclosure))
```

```{r}
sociodemographic_predictors = sociodemographic_predictors[!grepl('interview_date_numeric', sociodemographic_predictors)]
```

# Supplemental Analysis Adding a School Closure Missingness Predictor (to account for autocorrelation of missingness)

```{r}
m_formula_template_gps_subgroup = as.formula(paste('missing_gps ~ scaled_predictor*Control + 
                                               personal_time_z + weekend + summerbreak +', 
                                          paste(sociodemographic_predictors, collapse = '+'), ' + (1 |ID)'))

m_formula_template_accel_subgroup = as.formula(paste('missing_accel ~ scaled_predictor*Control + 
                                                 personal_time_z + weekend + summerbreak +', 
                                          paste(sociodemographic_predictors, collapse = '+'), ' + (1 |ID)'))

m_formula_template_keyboard_subgroup = as.formula(paste('missing_keyboard ~ scaled_predictor*Control +  
                                                    personal_time_z + weekend + summerbreak +', 
                                          paste(sociodemographic_predictors, collapse = '+'), ' + (1 |ID)'))

m_formula_template_daily_subgroup = as.formula(paste('missing_daily ~ scaled_predictor*Control + 
                                                           personal_time_z + weekend + summerbreak +', 
                                            paste(sociodemographic_predictors, collapse = '+'), ' + (1 |ID)'))

m_formula_template_risk_subgroup = as.formula(paste('risk_missing ~ scaled_predictor*Control + 
                                                personal_time_z + summerbreak +', 
                                           paste(sociodemographic_predictors, collapse = '+'), ' + (1 |ID)'))
```


# Run models & get ORs for each outcome

```{r risk_models}
risk_outputs = coef_pipeline(df = risk_missing, 
                      model_formula_template = m_formula_template_risk_subgroup,
                      clinical_predictors = clinical_predictors,
                      conf_level_corrected = corrected_conf_level,
                      bayes=TRUE)

# Save and clear unused memory
risk_outputs = risk_outputs[names(risk_outputs) != 'models_df']
save(risk_outputs, file = 'model_outputs/baseline_predictors_missingness_brm_risk_subgroup.rda')
gc()
```

```{r daily_models}
daily_outputs = coef_pipeline(df = daily, 
                      model_formula_template = m_formula_template_daily_subgroup,
                      clinical_predictors = clinical_predictors,
                      conf_level_corrected = corrected_conf_level,
                      bayes=TRUE)

# Save and clear unused memory
daily_outputs = daily_outputs[names(daily_outputs) != 'models_df']
save(daily_outputs, file = 'model_outputs/baseline_predictors_missingness_brm_daily_subgroup.rda')
gc()
```

```{r gps_models}
gps_outputs = coef_pipeline(df = daily, 
                      model_formula_template = m_formula_template_gps_subgroup,
                      clinical_predictors = clinical_predictors,
                      conf_level_corrected = corrected_conf_level,
                      bayes=TRUE)

# Save and clear unused memory
gps_outputs = gps_outputs[names(gps_outputs) != 'models_df']
save(gps_outputs, file = 'model_outputs/baseline_predictors_missingness_brm_gps_subgroup.rda')
gc()
```

```{r accel_models}
accel_outputs = coef_pipeline(df = daily, 
                      model_formula_template = m_formula_template_accel_subgroup,
                      clinical_predictors = clinical_predictors,
                      conf_level_corrected = corrected_conf_level,
                      bayes=TRUE)

# Save and clear unused memory
accel_outputs = accel_outputs[names(accel_outputs) != 'models_df']
save(accel_outputs, file = 'model_outputs/baseline_predictors_missingness_brm_accel_subgroup.rda')
gc()
```


```{r keyboard_models}
keyboard_outputs = coef_pipeline(df = daily, 
                      model_formula_template = m_formula_template_keyboard_subgroup,
                      clinical_predictors = clinical_predictors,
                      conf_level_corrected = corrected_conf_level,
                      bayes=TRUE)

keyboard_outputs = keyboard_outputs[names(keyboard_outputs) != 'models_df']
save(keyboard_outputs, file = 'model_outputs/baseline_predictors_missingness_brm_keyboard_subgroup.rda')
gc()
```

```{r}
load('model_outputs/baseline_predictors_missingness_brm_accel_subgroup.rda')
load('model_outputs/baseline_predictors_missingness_brm_gps_subgroup.rda')
load('model_outputs/baseline_predictors_missingness_brm_keyboard_subgroup.rda')
load('model_outputs/baseline_predictors_missingness_brm_daily_subgroup.rda')
load('model_outputs/baseline_predictors_missingness_brm_risk_subgroup.rda')


combined_coef_bayes = rbind(risk_outputs$coef_df %>% mutate(type = 'Weekly SI Survey'),
                    daily_outputs$coef_df %>% mutate(type = 'Daily Mood Survey'),
                    gps_outputs$coef_df %>% mutate(type = 'GPS'),
                    accel_outputs$coef_df %>% mutate(type = 'Accelerometer'),
                    keyboard_outputs$coef_df %>% mutate(type = 'Keyboard Input'))


g= risk_outputs$coef_df 
```

```{r}
combined_coef_bayes_interaction = dplyr::filter(combined_coef_bayes, term == 'scaled_predictor:Control1')

combined_coef_bayes_interaction = combined_coef_bayes_interaction %>%
    mutate(variable_category = case_when(
      name %in% c('current_medication', 'current_therapy', 'hospitalization_pastyear', 
                      'visit_er_pastyear') ~ 'Service',
      name %in% c('SITBI_SB_ever', 'SITBI_sa_ever', 'SITBI_engaged_nssi_ever', 
                      'SITBI_si_ever') ~ 'SITBI',
      name %in% c('SHAPS_TOT_r_i', 'MFQ_i', 'SCARED_i', 'SSI19_i', 'INQ_burdenensomeness_i', 
                      'INQ_thwartedbelonging_i', 'comorbidity_current') ~ 'Psychopathology'),
      name = dplyr::recode(name, 
                           'interview_age_years'='Age (Years)',
                           'SITBI_engaged_nssi_ever'='Lifetime NSSI (SITBI)',
                           'MFQ_i'='Depression Symptoms (MFQ)',
                           'SITBI_sa_ever'='Lifetime Suicide Attempt (SITBI)', 
                           'SHAPS_TOT_r_i'='Anhedonia Symptoms (SHAPS)',
                           'comorbidity_current'='# Current Diagnoses\n(MINI-KID)',
                           'SITBI_si_ever'='Lifetime Suicidal Ideation (SITBI)', 
                           'hospitalization_pastyear' = 'Past Year Hospitalization',
                           'visit_er_pastyear'='Past Year ED Visit', 
                           'SITBI_SB_ever'='Lifetime Suicidal Behavior (SITBI)',
                           'SCARED_i'='Anxiety Symptoms (SCARED)', 
                           'SSI19_i'='Suicidal Ideation Severity (SSI)',
                           'INQ_burdenensomeness_i'='Perceived Burdensomeness (INQ)', 
                           'INQ_thwartedbelonging_i'='Thwarted Belonging (INQ)',
                           'hispanic'='Hispanic', 'sexMTRUE'='Assigned Male Sex',
                           'siteCUMC'='Site (NYC)', 
                           'current_therapy'='Current Therapy',
                           'current_medication'='Current Medication'),
      continuous = ifelse(variable_category == 'Psychopathology', '*', ''),
      name = paste0(name, continuous),
      missingness_type = ifelse(grepl('Survey', type), 'Survey', 'Passive'))
  
    sig_frame = combined_coef_bayes_interaction %>%
      ungroup() %>%
      mutate(., sig_star = case_when(
        conf.low < 0 & conf.high <0 & coef_type == 'corrected' ~ '#',
        conf.low > 0 & conf.high >0 & coef_type == 'corrected' ~ '#',
        conf.low < 0 & conf.high <0 & coef_type == 'uncorrected' ~ '+',
        conf.low > 0 & conf.high >0 & coef_type == 'uncorrected' ~ '+',
        conf.low < 0 & conf.high >=0 ~ ' ',
        conf.low <=0 & conf.high >0 ~ ' '))

    sig_frame = pivot_wider(sig_frame,
                            id_cols = c('type', 'name', 'term'),
                            names_from = coef_type,
                            values_from = sig_star)

    sig_frame = mutate(sig_frame, sig_star = case_when(
      corrected == '#' ~ corrected,
      corrected != '#' ~ uncorrected
    ))
    
    
    sig_frame = left_join(combined_coef_bayes_interaction, sig_frame, by = c('type', 'name', 'term')) %>%
      dplyr::filter(term == 'scaled_predictor:Control1')
  

  # filter only for predictors of interest
  #or_df = dplyr::filter(or_df, term == 'scaled_predictor')
  
  # Create plot
  mlm_or_plt = combined_coef_bayes_interaction %>%
    dplyr::filter(!grepl('Suicid', name)) %>%
    dplyr::filter(coef_type == 'uncorrected') %>%
    ggplot(data=., aes(x = name, y = estimate)) +
    geom_point(position = position_dodge(0.6), aes(color = type)) +
      geom_errorbar(data = dplyr::filter(combined_coef_bayes_interaction, coef_type == 'corrected', !grepl('Suicid', name)),
                  aes(ymin = conf.low, ymax = conf.high, color = type),
                 width = 0, position = position_dodge(0.6)) +
      geom_errorbar(aes(ymin = conf.low, ymax = conf.high, color = type), 
                    lwd = 1, width = 0, position = position_dodge(0.6)) +
      geom_text(data = dplyr::filter(sig_frame, coef_type == 'corrected', !grepl('Suicid', name)), 
                aes(x = name, y = conf.high*1.1, label = sig_star, group = type, color = type),
                position = position_dodge(0.6), show.legend = FALSE) +
      coord_flip() +
      labs(x = NULL, y = 'Beta Coefficient for Group X Baseline Predictor Interactions\nValues >0 indicate a more positive association with missing data among psychiatric controls') +  
      geom_hline(yintercept = 0, lty = 2) +
      labs(color = NULL) +
      scale_color_manual(values= c(
                           "Accelerometer" = pal2[1], 
                           "Daily Mood Survey" = pal2[2], 
                           "GPS" = pal2[3], 
                           "Keyboard Input" = pal2[4], 
                           "Weekly SI Survey" = pal2[5]
                         ),
                         limits= c(
                           "Keyboard Input", 
                           "GPS", 
                           "Accelerometer",
                           "Weekly SI Survey",
                           "Daily Mood Survey"
                         )) +
      facet_grid(rows=vars(variable_category), cols = vars(missingness_type), drop = TRUE, scales = 'free_y', space = 'free_y') +
      scale_x_discrete(limits = rev) +
      theme(strip.text.y = element_blank())

ggsave(mlm_or_plt, file = '../../figures/supplement/clinical_baseline_predictors_subgroup_brm.png', height = 10, width =8)

```
